% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGB_LambertUng.R
\name{AGB_LambertUngDBH}
\alias{AGB_LambertUngDBH}
\title{Calculate tree-level AGB using Canadian national tree aboveground biomass equations using tree DBH as input}
\usage{
AGB_LambertUngDBH(DBH, species = NA)
}
\arguments{
\item{DBH}{Tree DBH}

\item{species}{Tree species code in the NFI standard (e.g. POPU.TRE)}
}
\value{
A list containing aboveground biomass values for tree components: wood, bark, stem, foliage, branches, crown, 
as well as the total tree aboveground biomass.
}
\description{
Calculates tree-level aboveground biomass [kg] using methods presented in Lambert et al. (2005) and Ung et al. (2008).
}
\examples{

AGB_LambertUngDBH(20)
AGB_LambertUngDBH(20, species="PINU.CON")
AGB_LambertUngDBH(10:30, species="PINU.CON")

}
\references{
Lambert, M. C., Ung, C. H., & Raulier, F. (2005). Canadian national tree aboveground biomass equations. Canadian Journal of Forest Research, 35(8), 1996–2018. https://doi.org/10.1139/x05-112

Ung, C.-H., Bernier, P., & Guo, X.-J. (2008). Canadian national biomass equations: new parameter estimates that include British Columbia data. Canadian Journal of Forest Research, 38(5), 1123–1132. https://doi.org/10.1139/X07-224
}
