% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AlbertaNaturalRegSubreg}
\alias{AlbertaNaturalRegSubreg}
\title{Natural regions and subregions of Alberta}
\format{
A data frame with four variables:
\describe{
  \item{NaturalRegion}{Natural region name}
  \item{NaturalSubregion}{Natural subregion name}
  \item{NaturalSubregionNum}{Subregion numeric code}
  \item{NaturalSubregionCode}{Subregion text code}
}
}
\usage{
AlbertaNaturalRegSubreg
}
\description{
A dataset containing a list of Alberta regions, subregions, and their codes.
}
\references{
https://open.alberta.ca/publications/0778545725
}
\keyword{datasets}
