% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CodesEcozones}
\alias{CodesEcozones}
\title{Ecozones of Canada and respective codes}
\format{
A data frame with three variables:
\describe{
  \item{EcozoneCode}{Ecozone code}
  \item{EcozoneName}{Ecozone name}
  \item{EcozoneNom}{Ecozone name in french}
}
}
\usage{
CodesEcozones
}
\description{
A dataset containing a list of Canada's ecozones and their codes.
}
\references{
https://open.canada.ca/data/en/dataset/7ad7ea01-eb23-4824-bccc-66adb7c5bdf8
}
\keyword{datasets}
