% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{merchcrit}
\alias{merchcrit}
\title{Merchantability criteria}
\format{
A data frame consisting of merchantability criteria.
}
\usage{
merchcrit
}
\description{
A dataset containing merchantability criteria, by jurisdiction, and species
}
\keyword{datasets}
