% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parameters_GalBella94}
\alias{parameters_GalBella94}
\title{Taper model parameters for Saskatchewan tree species}
\format{
A tibble with 12 rows and 12 columns:
\describe{
  \item{Species}{Species code following the Canadian National Forest Inventory (NFI) convention (e.g., \code{PICE.GLA}).}
  \item{a0, a1, a2}{Scaling and conditioning parameters of the K2 taper model.}
  \item{b1, b2, b3, b4, b5}{Exponent parameters controlling stem shape in the variable-exponent formulation.}
  \item{N}{Number of trees used to fit the model for the given species.}
  \item{RSS}{Residual sum of squares from the nonlinear model fit.}
}
}
\source{
Gal, J., & Bella, I.E. (1994). \emph{New stem taper functions for 12 Saskatchewan
timber species}. Natural Resources Canada, Canadian Forest Service, Northwest
Region, Information Report NOR-X-338. Table 5.
}
\usage{
parameters_GalBella94
}
\description{
Species-specific parameter estimates for the Kozak (1988) variable-exponent
taper model, as fitted by Gal & Bella (1994) using stem section data from
Saskatchewan. These parameters are intended for use in numerical stem taper
and volume calculations based on diameter at breast height (DBH) and total
tree height.
}
\details{
The dataset corresponds to Table 5 in Gal & Bella (1994) and includes parameter
estimates for 12 commercially important Saskatchewan tree species.


These parameters are valid for Saskatchewan conditions and should not be
applied outside the region without caution.
}
\references{
Kozak, A. (1988). A variable-exponent taper equation.
\emph{Canadian Journal of Forest Research}, 18, 1363–1368.
}
\keyword{datasets}
