% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parameters_Huang94}
\alias{parameters_Huang94}
\title{Taper model parameters for Alberta tree species}
\format{
A data frame:
\describe{
  \item{Species}{Tree species code following the NFI standard (e.g. POPU.TRE)}
  \item{Subregion}{Natural subregion of Alberta. 'Province' indicates provincial-level parameters (See \code{\link{AlbertaNaturalRegSubreg}} and Appendix 4 in Huang et al. 1994)}
  \item{a0, a1, a2}{Scaling and conditioning parameters of the K2 taper model.}
  \item{b1, b2, b3, b4, b5}{Exponent parameters controlling stem shape in the variable-exponent formulation.}
 }
}
\usage{
parameters_Huang94
}
\description{
Species-specific parameter estimates for the individual-tree volume models
developed by Huang (1994) for Alberta forests, designed to be used with Kozak (1988)
variable exponent taper model.
}
\details{
The dataset includes parameter sets applicable at the provincial level as
well as, where available, subregion-specific coefficients.
}
\references{
Huang, S. (1994). Ecologically Based Individual Tree Volume Estimation for Major Alberta Tree Species. Report 1 - Individual tree volume estimation procedures for Alberta: Methods of Formulation and Statistical Foundations. Alberta Environmental Protection, Land and Forest Service, Forest Management Division, Edmonton, AB.
}
\keyword{datasets}
