% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parameters_HuangV}
\alias{parameters_HuangV}
\title{Model parameters required for tree volume equations for major Alberta tree species}
\format{
A data frame with four variables:
\describe{
  \item{species}{Tree species code following the NFI standard (e.g. POPU.TRE)}
  \item{parameter}{Name of model parameter}
  \item{estimate}{Parameter estimate}
  \item{NaturalSubregionNum}{Number for natural subregions of Alberta (See \code{\link{AlbertaNaturalRegSubreg}} and Appendix 4 in Huang et al. 1994)}
  \item{NaturalSubregionCode}{Code for natural subregions of Alberta. 'Province' indicates provincial-level parameters (See \code{\link{AlbertaNaturalRegSubreg}} and Appendix 4 in Huang et al. 1994)}
}
}
\usage{
parameters_HuangV
}
\description{
A dataset containing parameters for the Huang et al 1994 tree volume models in Alberta
}
\references{
Huang, S., Titus, S., Lakusta, S., & Held, R. (1994). Ecologically Based Individual Tree Volume Estimation for Major Alberta Tree Species. In … to Alberta Environmental Protection, Land …. http://scholar.google.com/scholar?hl=en&btnG=Search&q=intitle:Ecologically+Based+Individual+Tree+Volume+Estimation+for+Major+Alberta+Tree+Species#0
}
\keyword{datasets}
