% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parameters_Kozak94}
\alias{parameters_Kozak94}
\title{Kozak (1994) BEC-zone taper model parameters (BC)}
\format{
A data frame containing species- and BEC-zone-specific parameter sets
  required by the Kozak BC taper implementation in this package.
}
\source{
Kozak, A. (1994). *Development of Taper Equations by BEC Zones and
  Species*. Province of British Columbia, Ministry of Forests (report).
}
\usage{
parameters_Kozak94
}
\description{
Model parameter table for the BC Ministry of Forests taper equations developed
by BEC zone and species (commonly referenced as the "Kozak 1994/1995" BC taper
parameters in inventory applications).
}
\keyword{datasets}
