% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parameters_LambertUng}
\alias{parameters_LambertUng}
\title{Model parameters required for Canadian national tree aboveground biomass equations}
\format{
A data frame with five variables:
\describe{
  \item{species}{Tree species code following the NFI standard (e.g. POPU.TRE)}
  \item{model}{Type of the AGB model - based on DBH-only, or DBH and height}
  \item{parameter}{Name of model parameter}
  \item{estimate}{Parameter estimate}
  \item{stderr}{Standard error of the estimate}
}
}
\usage{
parameters_LambertUng
}
\description{
A dataset containing the coefficients for the Lambert et al 2005 and Ung et al 2008 AGB models.
Parameters for both types of models are included (models based on DBH only and models based on DBH and height)
}
\references{
Lambert, M. C., Ung, C. H., & Raulier, F. (2005). Canadian national tree aboveground biomass equations. Canadian Journal of Forest Research, 35(8), 1996–2018. https://doi.org/10.1139/x05-112

Ung, C.-H., Bernier, P., & Guo, X.-J. (2008). Canadian national biomass equations: new parameter estimates that include British Columbia data. Canadian Journal of Forest Research, 38(5), 1123–1132. https://doi.org/10.1139/X07-224
}
\seealso{
\code{\link{AGB_LambertUngDBH}} and \code{\link{AGB_LambertUngDBHHT}}
}
\keyword{datasets}
