% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parameters_NationalTaperModelsDBH}
\alias{parameters_NationalTaperModelsDBH}
\title{Model parameters required for calculating volume using the taper model based on DBH}
\format{
An object of class \code{data.frame} with 34 rows and 9 columns.
}
\usage{
parameters_NationalTaperModelsDBH
}
\description{
A dataset containing parameters for the Ung et al 2013 models.
}
\details{
' @format A data frame consisting of model parameters. Table 6 in Ung et al 2013.
}
\references{
Ung, C.H., Guo, X.J., Fortin, M., 2013. Canadian national taper models. Forestry Chronicle 89, 211–224. https://doi.org/10.5558/tfc2013-040
}
\keyword{datasets}
