% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parameters_volNL}
\alias{parameters_volNL}
\title{Parameters for the Newfoundland and Labrador tree volume equations}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 46 rows and 14 columns.
}
\usage{
parameters_volNL
}
\description{
Parameters for the Newfoundland and Labrador tree volume equations
}
\keyword{datasets}
