% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{params_Vtot2Vmerch}
\alias{params_Vtot2Vmerch}
\title{Model parameters required for total volume to merchantable volume conversion}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2755 rows and 9 columns.
}
\usage{
params_Vtot2Vmerch
}
\description{
A dataset containing parameters for the Boudewyn et al 2007 total volume to merchantable volume conversion models. An updated version
of the parameters was downloaded from https://nfi.nfis.org/en/biomass_models.
}
\details{
' @format A data frame consisting of model parameters. Table 14 in Boudewyn et al 2007.
}
\references{
Boudewyn, P.A.; Song, X.; Magnussen, S.; Gillis, M.D. (2007). Model-based, volume-to-biomass conversion for forested and vegetated land in Canada. Natural Resources Canada, Canadian Forest Service, Pacific Forestry Centre, Victoria, BC. Information Report BC-X-411. 112 p.
}
\keyword{datasets}
