% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_model_registry.R
\name{si_model_registry_species}
\alias{si_model_registry_species}
\title{Site-index model registry with species coverage}
\usage{
si_model_registry_species()
}
\value{
A tibble like \code{si_model_registry()} plus:
\itemize{
\item \code{species} (list-column of character vectors)
\item \code{n_species} (integer)
\item \code{species_text} (collapsed string for printing)
}
}
\description{
Site-index model registry with species coverage
}
\keyword{internal}
