% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize_jurisdiction_code}
\alias{standardize_jurisdiction_code}
\title{Standardize jurisdiction (province and territorial) codes}
\usage{
standardize_jurisdiction_code(x)
}
\description{
Standardize jurisdiction (province and territorial) codes
}
\keyword{internal}
