% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize_species_code}
\alias{standardize_species_code}
\title{Standardize NFI species codes}
\usage{
standardize_species_code(x, keep_all = TRUE)
}
\description{
Converts compact codes like "PICEGLA" to "PICE.GLA". Keeps "ALL" unchanged.
}
\keyword{internal}
