% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_Huang.R
\name{V_Huang}
\alias{V_Huang}
\title{Calculate individual tree volume for major Alberta tree species}
\usage{
V_Huang(DBH, height, species, subregion = "Province")
}
\arguments{
\item{DBH}{Tree DBH}

\item{height}{Tree height}

\item{species}{Tree species code in the NFI standard (e.g. POPU.TRE)}

\item{subregion}{code depicting natural subregion of Alberta (e.g. "LH" for Lower Foothills). Please see
\code{\link{AlbertaNaturalRegSubreg}} for a complete list of Alberta natural regions and subregions. 
'Province' (the default) indicates province-level parameters.}
}
\value{
A list containing merchantable and total volume values [m3].
}
\description{
Calculates tree-level merchantable and total volume using methods presented in Huang et al. (1994).
}
\examples{
V_Huang(20, 20, "PICE.GLA")
V_Huang(20, 20, "PICE.GLA", subregion = "CP")

}
\references{
Huang, S., Titus, S., Lakusta, S., & Held, R. (1994). Ecologically Based Individual Tree Volume Estimation for Major Alberta Tree Species. In … to Alberta Environmental Protection, Land …. http://scholar.google.com/scholar?hl=en&btnG=Search&q=intitle:Ecologically+Based+Individual+Tree+Volume+Estimation+for+Major+Alberta+Tree+Species#0
}
