% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol_kozak88.R
\name{vol_galbella94}
\alias{vol_galbella94}
\title{Estimate tree volume in Saskatchewan using Kozak (1988) form with Gal & Bella (1994) params}
\usage{
vol_galbella94(DBH, height, species)
}
\arguments{
\item{DBH}{Numeric vector. Diameter at breast height (cm).}

\item{height}{Numeric vector. Total height (m).}

\item{species}{Character vector. Species code (e.g., "PICE.MAR").}
}
\description{
Estimate tree volume in Saskatchewan using Kozak (1988) form with Gal & Bella (1994) params
}
\references{
Gal, J., & Bella, I.E. (1994). \emph{New stem taper functions for 12 Saskatchewan timber species}. Natural Resources Canada, Canadian Forest Service, Northwest Region, Information Report NOR-X-338. Table 5.
Kozak, A. (1988). A variable-exponent taper equation.\emph{Canadian Journal of Forest Research}, 18, 1363–1368.
}
