% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol_honer83.R
\name{vol_honer83}
\alias{vol_honer83}
\title{Honer stem volume using DBH and total height}
\usage{
vol_honer83(DBH, height, species)
}
\arguments{
\item{DBH}{Tree diameter at breast height (cm)}

\item{height}{Total height (m)}

\item{species}{Species code (e.g., "PINU.CON")}
}
\value{
data.frame with vol_total and vol_merchantable
}
\description{
Estimates total and merchantable stem volume using the
Honer et al. (1983) model, developed for central and eastern Canada.
}
\examples{
# Single tree
vol_honer83(
  DBH = 30,
  height = 22,
  species = "PICE.GLA"
)

# Multiple trees, mixed species
vol_honer83(
  DBH = c(25, 32, 18),
  height = c(20, 24, 15),
  species = c("PICE.GLA", "ABIE.BAL", "PINU.BAN")
)

# Works with the pipe operator:
trees <- tibble::tibble(
  tree_id = 1:4,
  dbh = c(22, 30, 18, 35),
  ht  = c(18, 24, 15, 27),
  species = c("PICE.GLA", "PICE.GLA", "ABIE.BAL", "PINU.BAN")
)
trees |> dplyr::mutate(vol=vol_honer83(dbh, ht, species)) |> tidyr::unnest(vol)
}
\references{
Honer, T.G.; Ker, M.F.; Alemdag, I.S. 1983. Metric timber tables for the commercial tree species of central and eastern Canada.
Environ. Can., Can. For. Serv., Maritimes For. Res. Cent., Fredericton, NB. Inf. Rep. M-X-140.
}
