% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol_kozak88.R
\name{vol_huang94}
\alias{vol_huang94}
\title{Estimate tree volume in Alberta using Kozak (1988) model form with Huang (1994) parameters}
\usage{
vol_huang94(DBH, height, species, subregion = "Province")
}
\arguments{
\item{DBH}{Numeric vector. Diameter at breast height (cm).}

\item{height}{Numeric vector. Total height (m).}

\item{species}{Character vector. Species code (e.g., "PICE.MAR").}

\item{subregion}{Character vector. Alberta subregion name; defaults to "Province".}
}
\value{
A tibble with volumes (m^3): total and merchantable.
}
\description{
Estimate tree volume in Alberta using Kozak (1988) model form with Huang (1994) parameters
}
\references{
Huang, S. (1994). Ecologically Based Individual Tree Volume Estimation for Major Alberta Tree Species. Report 1 - Individual tree volume estimation procedures for Alberta: Methods of Formulation and Statistical Foundations. Alberta Environmental Protection, Land and Forest Service, Forest Management Division, Edmonton, AB.

Kozak, A. (1988). A variable-exponent taper equation. \emph{Canadian Journal of Forest Research}, 18, 1363–1368.
}
