% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol_kozak88.R
\name{vol_klos2004}
\alias{vol_klos2004}
\title{Estimate tree volume (Manitoba) using Kozak (1988) form with Klos (2004) params}
\usage{
vol_klos2004(DBH, height, species, subregion = "Province")
}
\arguments{
\item{DBH}{Numeric vector. Diameter at breast height (cm).}

\item{height}{Numeric vector. Total height (m).}

\item{species}{Character vector. Species code (e.g., "PICE.MAR").}

\item{subregion}{Optional Ecozone or Ecoregion identifier}
}
\description{
Estimate tree volume (Manitoba) using Kozak (1988) form with Klos (2004) params
}
\references{
Klos, R. J. (2004). \emph{Stem taper and volume equations for major tree species in Manitoba}. M.Sc. thesis, University of Manitoba, Winnipeg, MB, Canada.

Kozak, A. (1988). A variable-exponent taper equation. \emph{Canadian Journal of Forest Research}, 18, 1363–1368.
}
