% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol_kozak88.R
\name{vol_klos2007}
\alias{vol_klos2007}
\title{Estimate tree volume in Manitoba using Kozak (1988) form with Klos (2004) parameters}
\usage{
vol_klos2007(DBH, height, species, subregion = "Province")
}
\arguments{
\item{DBH}{Numeric vector. Diameter at breast height (cm).}

\item{height}{Numeric vector. Total height (m).}

\item{species}{Character vector. Species code (e.g., "PICE.MAR").}

\item{subregion}{Optional Ecozone or Ecoregion identifier}
}
\description{
Estimate tree volume in Manitoba using Kozak (1988) form with Klos (2004) parameters
}
\references{
Klos, R. J., Wang, G. G., Dang, Q.-L., & East, E. W. (2007).
\emph{Taper equations for five major commercial tree species in Manitoba, Canada}.
Western Journal of Applied Forestry, 22(3), 163–170.

Kozak, A. (1988). A variable-exponent taper equation. \emph{Canadian Journal of Forest Research}, 18, 1363–1368.
}
