% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol_ung_national.R
\name{vol_national_dbh}
\alias{vol_national_dbh}
\title{Canadian national taper model (DBH only)}
\usage{
vol_national_dbh(DBH, species, jurisdiction)
}
\arguments{
\item{DBH}{Numeric vector of DBH (cm).}

\item{species}{Character. Species code (e.g., "PICE.GLA").}

\item{jurisdiction}{Character. Jurisdiction code (e.g., "BC", "ON").}
}
\value{
A tibble with columns `vol_merchantable` and `vol_total`.
}
\description{
Canadian national taper model (DBH only)
}
\references{
Ung, C.H., Guo, X.J., Fortin, M., 2013. Canadian national taper models. Forestry Chronicle 89, 211–224.
https://doi.org/10.5558/tfc2013-040
}
