% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol_ung_national.R
\name{vol_national_dbh_ht}
\alias{vol_national_dbh_ht}
\title{Canadian national taper model (DBH and total height)}
\usage{
vol_national_dbh_ht(DBH, height, species, jurisdiction)
}
\arguments{
\item{DBH}{Numeric vector of DBH (cm).}

\item{species}{Character. Species code (e.g., "PICE.GLA").}

\item{jurisdiction}{Character. Jurisdiction code (e.g., "BC", "ON").}

\item{H}{Numeric vector of total height (m). Must be same length as DBH,
or length 1 (recycled).}
}
\value{
A tibble with columns `vol_merchantable` and `vol_total`.
}
\description{
Canadian national taper model (DBH and total height)
}
\references{
Ung, C.H., Guo, X.J., Fortin, M., 2013. Canadian national taper models. Forestry Chronicle 89, 211–224.
https://doi.org/10.5558/tfc2013-040
}
