% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol_zakrzewski2013.R
\name{vol_zakrzewski2013}
\alias{vol_zakrzewski2013}
\title{Estimate tree volume using Zakrzewski (2013) taper model (Ontario)}
\source{
Zakrzewski, W. T., Penner, M. (2013). \emph{A Comparison of Tree Stem Taper Models
for Use in Ontario}. Ontario Forest Research Institute, Report 176.
}
\usage{
vol_zakrzewski2013(DBH, height, species)
}
\arguments{
\item{DBH}{Numeric vector of diameter at breast height (cm).}

\item{height}{Numeric vector of total tree height (m).}

\item{species}{Character vector of species codes}
}
\value{
A tibble with volumes (m^3): total and merchantable.
}
\description{
Estimate tree volume using Zakrzewski (2013) taper model (Ontario)
}
\examples{
# Single tree
vol_zakrzewski2013(
  DBH = 20,
  height = 22,
  species = "PICE.MAR"
)

# Vectorized input (multiple trees)
vol_zakrzewski2013(
  DBH = c(20, 30, 40),
  height = c(18, 22, 28),
  species = c("PICE.MAR", "PINU.STR", "THUJ.OCC")
)

}
