% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_model_registry.R
\name{volume_model_registry_species}
\alias{volume_model_registry_species}
\title{Volume model registry with species coverage}
\usage{
volume_model_registry_species()
}
\value{
A tibble like \code{volume_model_registry()} plus:
\itemize{
\item \code{species} (list-column of character vectors)
\item \code{n_species} (integer)
\item \code{species_text} (collapsed string for printing)
}
}
\description{
Volume model registry with species coverage
}
\keyword{internal}
