% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vtot2Vmerch.R
\name{Vtot2Vmerch}
\alias{Vtot2Vmerch}
\title{Total volume to gross (net in B.C.) merchantable volume conversion}
\usage{
Vtot2Vmerch(total_volume, species, jurisdiction, ecozone)
}
\arguments{
\item{total_volume}{Gross merchantable volume/ha (net in BC) of all live trees}

\item{species}{Species code in the NFI standard (e.g. POPU.TRE)}

\item{jurisdiction}{A two-letter code depicting jurisdiction (e.g. "AB", "BC", etc)}

\item{ecozone}{ecozone number (1-15). Call 'CodesEcozones' for a list of ecozones names and codes - also available in table 2 of appendix 7 of Boudewyn et al (2007).}
}
\value{
Merchantable volume
}
\description{
Total volume to gross (net in B.C.) merchantable volume conversion based on Appendix 6 in Boudewyn et al. 2007 (see references).
}
\examples{
Vtot2Vmerch(total_volume = 300, species = "PINU.CON", jurisdiction = "AB", ecozone = 4)

}
\references{
Boudewyn, P.A.; Song, X.; Magnussen, S.; Gillis, M.D. (2007). Model-based, volume-to-biomass conversion for forested and vegetated land in Canada. Natural Resources Canada, Canadian Forest Service, Pacific Forestry Centre, Victoria, BC. Information Report BC-X-411. 112 p.
}
