% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytbh_nigh2000.R
\name{ytbh_nigh2000}
\alias{ytbh_nigh2000}
\title{Nigh (2000) years-to-breast-height model for interior western redcedar}
\usage{
ytbh_nigh2000(si)
}
\arguments{
\item{si}{Numeric vector. Site index (m, base age 50 years at breast height).}
}
\value{
A tibble with one column:
\describe{
\item{ytbh}{Predicted years to breast height (years).}
}
}
\description{
Iimplementation of the Nigh (2000) years-to-breast-height model
for interior western redcedar (\code{THUJ.PLI}).
}
\details{
The source equation is:
\deqn{YTBH = 18.18 - 0.5526 \times SI}
}
\examples{
ytbh_nigh2000(
  si = c(10, 15, 20)
)

}
\references{
Nigh, G. D. (2000). Western redcedar site index models for the interior of
British Columbia. British Columbia Ministry of Forests, Research Report 18.
}
