% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytbh_nighcourtin1998.R
\name{ytbh_nighcourtin1998}
\alias{ytbh_nighcourtin1998}
\title{Nigh and Courtin (1998) years-to-breast-height model for red alder}
\usage{
ytbh_nighcourtin1998(si)
}
\arguments{
\item{si}{Numeric vector. Site index SI25 (m at breast-height age 25 years).}
}
\value{
A tibble with one column:
\describe{
\item{ytbh}{Predicted years to breast height (years).}
}
}
\description{
Implementation of the Nigh and Courtin (1998) years-to-breast-height
model for red alder (\code{ALNU.RUB}) in coastal British Columbia.
}
\details{
The source relation is piecewise:
\deqn{
YTBH = 5.494 - 0.1789 \times SI25, \quad SI25 \le 25
}
\deqn{
YTBH = 1.0, \quad SI25 > 25
}
}
\examples{
ytbh_nighcourtin1998(
  si = c(12, 20, 28)
)

}
\references{
Nigh, G. D., & Courtin, P. J. (1998). Height models for Red Alder
(\emph{Alnus rubra} Bong.) in British Columbia.
\emph{New Forests}, 16, 59-70.
}
