% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ung2009.R
\name{Ung2009}
\alias{Ung2009}
\title{Ung2009 Growth and Yield Model}
\usage{
Ung2009(species, age, GDD, PREC)
}
\arguments{
\item{species}{Character. A valid species name as defined in the parameter table of the model.}

\item{age}{Numeric. The age of the stand in years. Must be greater than 0.}

\item{GDD}{Numeric. Growth degree days.}

\item{PREC}{Numeric. Precipitation.}
}
\value{
A list containing:
\describe{
  \item{H}{Predicted tree height (m).}
  \item{BA}{Predicted basal area (m²/ha).}
  \item{V}{Predicted volume (m³/ha).}
}
}
\description{
This function implements the model described in Ung et al. 2009.
}
\details{
The model predicts tree height (H), basal area (BA), and volume (V) based on 
species, age, growth degree days (GDD), and precipitation (PREC).
}
\examples{
# Example usage:
Ung2009(
  species = "ABIE.BAL",
  age = 1:150,
  GDD = 1500,
  PREC = 800
)

}
\references{
Ung, C.-H., Bernier, P.Y., Guo, X.J., Lambert, M.-C. (2009). 
"A simple growth and yield model for assessing changes in standing volume across Canada’s forests." 
The Forestry Chronicle, 85, 57–64. \doi{10.5558/tfc85057-1}.
}
