% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V2B.R
\name{V2B}
\alias{V2B}
\title{Volume-to-biomass conversion}
\usage{
V2B(volume, species, jurisdiction, ecozone)
}
\arguments{
\item{volume}{Gross merchantable volume/ha (net in BC) of all live trees}

\item{species}{Species code in the NFI standard (e.g. POPU.TRE)}

\item{jurisdiction}{A two-letter code depicting jurisdiction (e.g. "AB")}

\item{ecozone}{ecozone number (1-15)}
}
\value{
A list containing aboveground biomass values. Column names correspond to variable names
used in Boudewyn et al. 2007: 
\describe{
\item{\code{b_m}}{Total stem wood biomass of merchantable-sized live trees (biomass includes stumps and tops), in metric tonnes per ha.}
\item{\code{b_n}}{stem wood biomass of live, nonmerchantable-sized trees (tonnes/ha)}
\item{\code{b_mn}}{b_m + b_n}
\item{\code{b_s}}{stem wood biomass of live, sapling-sized trees (tonnes/ha)}
\item{\code{b_total}}{Total tree biomass}
\item{\code{b_bark}}{Total bark biomass}
\item{\code{b_branches}}{Total branch biomass}
\item{\code{b_foliage}}{Total foliage biomass}
}
}
\description{
Implementation of the model-based, volume-to-biomass conversion equations by Boudewyn et al. (2007).
Note - only scenarios 1 and 2 are currently implemented.
}
\examples{
V2B(350, species = "PINU.CON",jurisdiction = "BC", ecozone=4)

}
\references{
Boudewyn, P.A.; Song, X.; Magnussen, S.; Gillis, M.D. (2007). Model-based, volume-to-biomass conversion for forested and vegetated land in Canada. Natural Resources Canada, Canadian Forest Service, Pacific Forestry Centre, Victoria, BC. Information Report BC-X-411. 112 p.
}
