% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol_huang94.R
\name{vol_huang94}
\alias{vol_huang94}
\title{Calculate individual tree volume for major Alberta tree species}
\usage{
vol_huang94(DBH, height, species, subregion = "Province")
}
\arguments{
\item{DBH}{Tree DBH}

\item{height}{Tree height}

\item{species}{Tree species code in the NFI standard (e.g. POPU.TRE)}

\item{subregion}{code depicting natural subregion of Alberta (e.g. "LH" for Lower Foothills). Please see
\code{\link{AlbertaNaturalRegSubreg}} for a complete list of Alberta natural regions and subregions.
'Province' (the default) indicates province-level parameters.}
}
\value{
A tibble with volumes (m^3): total and merchantable.
}
\description{
Calculates tree-level merchantable and total volume using methods presented in Huang (1994).
}
\examples{
vol_huang94(20, 20, "PICE.GLA")
vol_huang94(20, 20, "PICE.GLA", subregion = "CP")

}
\references{
Huang, S. (1994). Ecologically Based Individual Tree Volume Estimation for Major Alberta Tree Species. Report 1 - Individual tree volume estimation procedures for Alberta: Methods of Formulation and Statistical Foundations. Alberta Environmental Protection, Land and Forest Service, Forest Management Division, Edmonton, AB.
}
